Consumidor/requirements.txt
requests~=2.28.2
kafka-python